formal file pats structure
